+++ io7m-smfj-frontend/src/main/java/com/io7m/smfj/frontend/SMFFCopier.java
@@ -291,8 +291,8 @@ public final class SMFFCopier implements SMFFCopierType
 
   @Override
   public boolean onMeta(
-    final int vendor,
-    final int schema,
+    final long vendor,
+    final long schema,
     final long length)
   {
     return true;
@@ -300,8 +300,8 @@ public final class SMFFCopier implements SMFFCopierType
 
   @Override
   public void onMetaData(
-    final int vendor,
-    final int schema,
+    final long vendor,
+    final long schema,
     final byte[] data)
   {
     try {
