+++ io7m-smfj-tests/src/test/java/com/io7m/smfj/tests/format/binary/FMB.java
@@ -79,8 +79,8 @@ public final class FMB
         {
           @Override
           public boolean onMeta(
-            final int vendor,
-            final int schema,
+            final long vendor,
+            final long schema,
             final long length)
           {
             return true;
@@ -88,14 +88,14 @@ public final class FMB
 
           @Override
           public void onMetaData(
-            final int vendor,
-            final int schema,
+            final long vendor,
+            final long schema,
             final byte[] data)
           {
             LOG.debug(
               "metadata: {} {} {}",
-              Integer.toUnsignedString(vendor, 16),
-              Integer.toUnsignedString(schema, 16),
+              Long.toUnsignedString(vendor, 16),
+              Long.toUnsignedString(schema, 16),
               Base64.getUrlEncoder().encodeToString(data));
           }
